Feature/test and update app manager

+++ HikiCoffee.Data/Entities/BillInfo.cs
@@ -10,9 +10,9 @@
 
         public int Quantity { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         public Bill Bill { get; set; }
