+++ HikiCoffee.ViewModels/BillInfos/BillInfoDataRequest/BillInfoCreateRequest.cs
@@ -8,8 +8,8 @@
 
         public int Quantity { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 }
