+++ HikiCoffee.ViewModels/Bills/BillViewModel.cs
@@ -16,7 +16,7 @@ namespace HikiCoffee.ViewModels.Bills
 
         public DateTime? DateCheckOut { get; set; }
 
-        public double TotalPayPrice { get; set; }
+        public decimal TotalPayPrice { get; set; }
 
         public List<BillInfoViewModel?> BillInfos { get; set; }
     }
